@@ -1,69 +1,57 @@
 {
   "name": "jecabeda-blog",
   "description": "José Cabeda",
-  "version": "4.1.0",
+  "version": "4.2.0",
   "author": "José Cabeda <jecabeda@gmail.com>",
   "bugs": {
     "url": "https://github.com/Cabeda/jecabeda-blog/issues"
   },
   "engines": {
-    "node": "14.x"
+    "node": "16.x"
   },
   "dependencies": {
-    "axios": "^0.26.1",
-    "babel-plugin-styled-components": "^2.0.7",
-    "babel-preset-react": "^6.24.1",
-    "dotenv": "^16.0.0",
-    "fontsource-merriweather": "^4.0.0",
-    "fontsource-montserrat": "^4.0.0",
-    "gatsby": "^4.11.2",
-    "gatsby-link": "^4.11.1",
-    "gatsby-plugin-feed": "^4.11.1",
-    "gatsby-plugin-image": "^2.11.1",
-    "gatsby-plugin-manifest": "^4.11.1",
-    "gatsby-plugin-offline": "^5.11.1",
-    "gatsby-plugin-postcss": "^5.11.0",
-    "gatsby-plugin-react-helmet": "^5.11.0",
-    "gatsby-plugin-sharp": "^4.11.1",
-    "gatsby-plugin-styled-components": "^5.11.0",
-    "gatsby-plugin-typography": "^4.11.0",
-    "gatsby-remark-copy-linked-files": "^5.11.0",
-    "gatsby-remark-images": "^6.11.1",
-    "gatsby-remark-prismjs": "^6.11.0",
-    "gatsby-remark-responsive-iframe": "^5.11.0",
-    "gatsby-remark-smartypants": "^5.11.0",
-    "gatsby-source-filesystem": "^4.11.1",
-    "gatsby-transformer-remark": "^5.11.1",
-    "gatsby-transformer-sharp": "^4.11.0",
+    "axios": "^0.27.2",
+    "dotenv": "^16.0.1",
+    "gatsby": "^4.20.0",
+    "gatsby-link": "^4.20.0",
+    "gatsby-plugin-feed": "^4.20.0",
+    "gatsby-plugin-image": "^2.20.0",
+    "gatsby-plugin-manifest": "^4.20.0",
+    "gatsby-plugin-offline": "^5.20.0",
+    "gatsby-plugin-postcss": "^5.20.0",
+    "gatsby-plugin-sharp": "^4.20.0",
+    "gatsby-plugin-styled-components": "^5.20.0",
+    "gatsby-remark-copy-linked-files": "^5.20.0",
+    "gatsby-remark-images": "^6.20.0",
+    "gatsby-remark-prismjs": "^6.20.0",
+    "gatsby-remark-responsive-iframe": "^5.20.0",
+    "gatsby-remark-smartypants": "^5.20.0",
+    "gatsby-source-filesystem": "^4.20.0",
+    "gatsby-transformer-remark": "^5.20.0",
+    "gatsby-transformer-sharp": "^4.20.0",
     "mysql2": "^2.3.3",
-    "postcss": "^8.4.12",
-    "prismjs": "^1.27.0",
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0",
-    "react-helmet": "^6.1.0",
-    "react-icons": "^4.3.1",
-    "react-typography": "^0.16.20",
+    "postcss": "^8.4.16",
+    "prismjs": "^1.28.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-icons": "^4.4.0",
+    "start": "^5.1.0",
     "styled-components": "^5.3.5",
-    "typography": "^0.16.21",
-    "typography-theme-wordpress-2016": "^0.16.19"
+    "typeface-merriweather": "0.0.72",
+    "typeface-montserrat": "0.0.75"
   },
   "devDependencies": {
     "@vercel/node": "^1.14.0",
-    "babel-eslint": "^10.1.0",
-    "babel-preset-gatsby": "^2.11.1",
     "cross-env": "^7.0.3",
-    "eslint": "^8.13.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jsx-a11y": "^6.5.1",
-    "eslint-plugin-react": "^7.29.4",
-    "eslint-plugin-react-hooks": "^4.4.0",
     "esm": "^3.2.25",
-    "prettier": "^2.6.2"
+    "prettier": "^2.7.1"
   },
   "homepage": "https://github.com/Cabeda/jecabeda-blog",
   "keywords": [
-    "gatsby"
+    "gatsby",
+    "dev",
+    "blog",
+    "cabeda"
   ],
   "license": "MIT",
   "main": "n/a",
