@@ -1,5 +1,4 @@
 import styled from "styled-components";
-import { rhythm } from '../utils/typography'
 
 const LayoutStyles = styled.div`
   body {
@@ -15,8 +14,8 @@ const LayoutStyles = styled.div`
 
   margin-left: auto;
   margin-right: auto;
-  max-width: ${rhythm(24)};
-  padding: ${rhythm(1.5)} ${rhythm(3 / 4)};
+  max-width: 42rem;
+  padding: 2.5rem 1.25rem;
 `;
 
 export default LayoutStyles;
